Squashed commit of the following:
commit f4904ade63324b668a0e21220413e1e12d0f6e12
Author: heejung0413 <acb4287@gmail.com>
Date:   Wed Mar 20 20:12:42 2024 +0900

    fix: chakraProvider 설정
